[Data] Treat one wav as one training example
Instead of each training example being the size of the receptive
field, which was done originally to save on memory, one forward backward
pass is one wav. The cross entropy loss of all samples are averaged across one wav.
This (obviously) trains by several of orders of magnitude faster.

Note: batchsize should be 1 right now, as there is no support for
padding and masking yet. Probably won't add this anyway because
the length of a wav means large memory usage.

Test plan:
1) th test/test_suite.lua
2) th main.lua -gpuids 1 -num_expblocks 3 -max_dilation 64 -dilated_nummaps 4,4,4,4,4,4,4 -onedim_nummaps 4 -output_nummaps 8 -method adam -lr 0.001 -split_dirname split_oneperson